move HandlerContext into its own file
