+++ sdks/go/pkg/catena/server.go
@@ -45,13 +45,11 @@ import (
 	"fmt"
 	"maps"
 	"math"
-	"slices"
 	"strconv"
 	"strings"
 	"sync"
 	"time"
 
-	"github.com/golang-jwt/jwt/v5"
 	"github.com/rossvideo/catena/sdks/go/pkg/config"
 	"github.com/rossvideo/catena/sdks/go/pkg/logger"
 	"github.com/rossvideo/catena/sdks/go/pkg/protos"
@@ -101,92 +99,6 @@ func (e EndpointType) String() string {
 		return "Connect"
 	default:
 		return "Unknown"
-	}
-}
-
-type TransportContext struct {
-	AccessToken string
-	Metadata    map[string][]string
-	Ctx         context.Context
-}
-
-// HandlerContext gives a handler everything it needs to know about the request
-// it is serving. Use it to:
-//   - check the caller's authorization with the scope helpers
-//     (HasReadScope/HasWriteScope/HasAnyReadScope/HasAnyWriteScope). When
-//     authorization is disabled these all return true.
-//   - read the request Metadata (for example, transport headers).
-//   - respect cancellation and deadlines through Context() (see below),
-//     which matters most for long-running or streaming handlers.
-//
-// Treat it as read-only; the server fills it in before your handler runs.
-type HandlerContext struct {
-	Token        *jwt.Token
-	readScopes   map[string]struct{}
-	writeScopes  map[string]struct{}
-	Metadata     map[string][]string
-	authzEnabled bool
-	ctx          context.Context
-	// ctxCancel tears down ctx and unregisters its shutdown watcher. The gate
-	// builds ctx and stashes cancel here; whoever owns the end of the request
-	// (invokeHandler's defer, GetSlots, or a torn-down streaming Connection)
-	// calls release(). Do not call this field directly - use release() so the
-	// zero HandlerContext (nil cancel) stays safe.
-	ctxCancel context.CancelFunc
-}
-
-// HasReadScope reports whether the caller was granted the named read scope.
-// A write grant implies read: holding "foo:w" satisfies HasReadScope("foo") as
-// well as HasWriteScope("foo"). It returns true for every scope when
-// authorization is disabled.
-func (ctx HandlerContext) HasReadScope(scopeName string) bool {
-	if !ctx.authzEnabled {
-		return true
-	}
-	_, ok := ctx.readScopes[scopeName]
-	return ok
-}
-
-// HasWriteScope reports whether the caller was granted the named write scope.
-// It returns true for every scope when authorization is disabled.
-func (ctx HandlerContext) HasWriteScope(scopeName string) bool {
-	if !ctx.authzEnabled {
-		return true
-	}
-	_, ok := ctx.writeScopes[scopeName]
-	return ok
-}
-
-// HasAnyWriteScope reports whether the caller holds at least one write scope,
-// i.e. whether they are allowed to make any change at all. Use it as a coarse
-// gate before the more specific HasWriteScope check.
-func (ctx HandlerContext) HasAnyWriteScope() bool {
-	return slices.ContainsFunc(catenaScopes, ctx.HasWriteScope)
-}
-
-// HasAnyReadScope reports whether the caller holds at least one read scope,
-// i.e. whether they are allowed to read anything at all. Use it as a coarse
-// gate before the more specific HasReadScope check.
-func (ctx HandlerContext) HasAnyReadScope() bool {
-	return slices.ContainsFunc(catenaScopes, ctx.HasReadScope)
-}
-
-// Context returns the context.Context for this request. It is Done when either
-// the request (or stream) ends or the server begins shutting down, whichever
-// happens first, and it carries any deadline and values from the request.
-// Long-running and streaming handlers should select on ctx.Done() (or pass ctx
-// to cancellable calls) so they stop promptly when the caller goes away or the
-// server is stopping.
-func (ctx HandlerContext) Context() context.Context {
-	return ctx.ctx
-}
-
-// release cancels the request context and unregisters its shutdown watcher. It
-// is safe to call multiple times and on a zero HandlerContext (nil ctxCancel),
-// so callers can defer it unconditionally after the gate.
-func (ctx HandlerContext) release() {
-	if ctx.ctxCancel != nil {
-		ctx.ctxCancel()
 	}
 }
